wpt: Fix event routing for clicking elements in child windows.
This change fixes the WPT infra for simulating click events on elements
in child windows with the following:

- In the renderer, use the window for the element instead of the current
Document's window when dispatching the event to the RWHV in the browser.

- In the browser, dispatch the event directly to the view's window
instead of the platform routing code. The latter routes the event to the
top level WC's window instead of the child WC's window.

Bug:1522213
Change-Id: I5e71c219528c12885ede682bb1d07f6c64c151bb