Shared Storage: Change to per-origin byte limit
To increase developer ergonomics, we drop the limit on number of
key-value pairs per origin, as well as the previous character limits
for keys and values. We move instead to a total byte quota for each
origin.

Note that, strictly speaking, there will still a key and value length
limit, in that each key and value must necessarily be less than or
equal to the quota, given that the total of all lengths of keys and
values must not exceed the quota.

We accomplish this by updating the previous feature param
`blink::features::kMaxSharedStorageEntriesPerOrigin` to a new
feature param `blink::features::kMaxSharedStorageBytesPerOrigin`,
deleting the previous feature param
`blink::features::kMaxSharedStorageStringLength`, and replacing
instances of the later with half of the new value for
`blink::features::kMaxSharedStorageBytesPerOrigin` (as we are
measuring the length in number of char16_t's).

For more information, see
https://github.com/WICG/shared-storage/pull/132.

Bug: 325352528,1218540
Change-Id: Icf3c390fcdb2893e74d1112c53db0ece2ecefe7b